Implemented Config Folder
The docker start script now copies the Cumulus.ini from the /config folder into the /opt/CumulusMX folder inside the container prior to starting, and copies it back when the application exits. This allows for improved handling of new installs, and moves the Config information out of the root directory. The config folder is now included in the Build script.